Flatten agents/prompts/ directory and rename metrics_interactive
